requirements.txt
pandas
scikit-learn
numpy
matplotlib
seaborn
psycopg2-binary
sqlalchemy
joblib
Flask 